+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Setup dist files.
 -   Icon color adapt according to the color theme.
 -   Add clipboard JS functions.
+-   Add pill option style for inputs.
+-   Add share JS function.
+-   Set rich links on head tag in html.
 
 ### Changed